add some casts
git-svn-id: https://svn.r-project.org/R/trunk@59115 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/parallel/src/rngstream.c
@@ -68,7 +68,7 @@ SEXP nextStream(SEXP x)
     }
     SEXP ans = allocVector(INTSXP, 7);
     INTEGER(ans)[0] = INTEGER(x)[0];
-    for (int i = 0;  i < 6; i++) INTEGER(ans)[i+1] = nseed[i];
+    for (int i = 0;  i < 6; i++) INTEGER(ans)[i+1] = (int) nseed[i];
     return ans;
 }
 
@@ -94,6 +94,6 @@ SEXP nextSubStream(SEXP x)
     }
     SEXP ans = allocVector(INTSXP, 7);
     INTEGER(ans)[0] = INTEGER(x)[0];
-    for (int i = 0;  i < 6; i++) INTEGER(ans)[i+1] = nseed[i];
+    for (int i = 0;  i < 6; i++) INTEGER(ans)[i+1] = (int) nseed[i];
     return ans;
 }
